pyproject.toml
[tool.poetry]
name = "log-parser"
version = "0.1.0"
description = ""
authors = ["alllexxx1 <alekkorpachyov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
django-filter = "^24.3"
pytest-django = "^4.9.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
